refactor(table): update table custom with table toolkit
