+++ src/shared/components/Table/Body/Line/__tests__/Line.test.tsx
@@ -13,17 +13,17 @@ const columnsMock = [
 
 describe('Line', () => {
   it.each`
-    columns        | className    | modifier      | children
+    columns        | className    | classModifier | children
     ${undefined}   | ${undefined} | ${undefined}  | ${undefined}
     ${[]}          | ${undefined} | ${undefined}  | ${undefined}
     ${columnsMock} | ${undefined} | ${undefined}  | ${undefined}
     ${columnsMock} | ${undefined} | ${'modifier'} | ${undefined}
     ${columnsMock} | ${'myclass'} | ${'modifier'} | ${(<td>child</td>)}
   `(
-    'Should render <Line/> when columns: $columns, className: $className, modifier: $modifier, children: $children',
-    ({ columns, className, modifier, children }) => {
+    'Should render <Line/> when columns: $columns, className: $className, classModifier: $classModifier, children: $children',
+    ({ columns, className, classModifier, children }) => {
       const { baseElement } = renderWithContainer(
-        <Line className={className} columns={columns} modifier={modifier}>
+        <Line className={className} columns={columns} classModifier={classModifier}>
           {children}
         </Line>,
         container,
