+++ src/shared/hoc/WithClassNameModifier.tsx
@@ -1,15 +1,24 @@
 import { ComponentType } from 'react';
 import { ClassManager } from '@axa-fr/react-toolkit-core';
 
-type TwithClassModifier = {
+export type TwithClassNameModifier = {
   className?: string;
   classModifier?: string;
   defaultClassName?: string;
 };
 
-const withClassNameModifier =
-  <P extends object>(Component: ComponentType<P>, classNameDefault = ''): ComponentType<P & TwithClassModifier> =>
-  ({ className = classNameDefault, classModifier = '', defaultClassName = '', ...props }: TwithClassModifier) =>
-    <Component {...(props as P)} className={ClassManager.getComponentClassName(className, classModifier, defaultClassName)} />;
+const withClassNameModifier = <P extends object>(
+  Component: ComponentType<P>,
+  defaultProps?: Partial<P>,
+  displayName = '',
+): ComponentType<P & TwithClassNameModifier> => {
+  const Hoc = ({ className = '', classModifier = '', defaultClassName = '', ...props }: TwithClassNameModifier & P) => (
+    <Component {...(props as P)} className={ClassManager.getComponentClassName(className, classModifier, defaultClassName)} />
+  );
+
+  Hoc.displayName = `WithClassModier(${Component.name || Component.displayName || displayName}`;
+  Hoc.defaultProps = defaultProps || {};
+  return Hoc;
+};
 
 export default withClassNameModifier;
